Moves the LRUcache to Controller level
